Use contenthash instead of hash
Using hash would generate a different hash when the JS bundle is built on each server instance. Thus, if we are using a load balancer, it's possible a server instance could receive a request for a hash that exists on a different instance, and in which case, the former would not be able to properly serve the JS file. Instead, it would return index.html, which is not what we want.